Extract Vercel deployment components
Move Vercel CLI arguments, output parsing, Docker digest parsing, project naming, Build Output generation, environment mapping, deployment model validation, plan writing, and state handling into focused internal components while keeping the deployment step as the Aspire pipeline coordinator.
