+++ src/CommunityToolkit.Aspire.Hosting.Vercel/VercelCliOutputParser.cs
@@ -1,0 +1,194 @@
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.Text.Json;
+using System.Text.RegularExpressions;
+using Aspire.Hosting;
+
+namespace CommunityToolkit.Aspire.Hosting.Vercel;
+
+internal static class VercelCliOutputParser
+{
+    public static string GetDeploymentUrl(string standardOutput)
+        => GetDeploymentResult(standardOutput).DeploymentUrl;
+
+    public static VercelDeploymentResult GetDeploymentResult(string standardOutput)
+    {
+        // `vercel deploy` output has changed between CLI versions and flags. Prefer structured
+        // JSON when present, then fall back to the last plain HTTP(S) URL printed by the CLI.
+        if (TryGetJsonDeploymentResult(standardOutput) is { } jsonDeploymentResult)
+        {
+            return jsonDeploymentResult;
+        }
+
+        // Older CLI versions printed the deployment URL as plain text. Keep the fallback so
+        // we fail only when no usable URL exists, not because formatting changed slightly.
+        string[] lines = standardOutput
+            .Split(['\r', '\n'], StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
+
+        string? deploymentUrl = lines.LastOrDefault(IsHttpUrl);
+        if (deploymentUrl is null)
+        {
+            throw new DistributedApplicationException($"Vercel deploy output did not contain an HTTP or HTTPS deployment URL. Output: {GetTrimmedOutput(standardOutput)}");
+        }
+
+        return new(DeploymentId: null, deploymentUrl);
+    }
+
+    public static VercelDeploymentInspection GetDeploymentInspection(string standardOutput)
+    {
+        try
+        {
+            // Parse the Vercel inspect JSON shapes observed across CLI versions:
+            //   { "readyState": "READY" }
+            //   { "state": "READY" }
+            //   { "deployment": { "readyState": "READY" } }
+            //   { "deployment": { "state": "READY" } }
+            using var document = JsonDocument.Parse(standardOutput);
+            var root = document.RootElement;
+            string? readyState = VercelJson.GetStringProperty(root, "readyState")
+                ?? VercelJson.GetStringProperty(root, "state")
+                ?? (root.TryGetProperty("deployment", out var deployment) ? VercelJson.GetStringProperty(deployment, "readyState") : null)
+                ?? (root.TryGetProperty("deployment", out deployment) ? VercelJson.GetStringProperty(deployment, "state") : null);
+
+            return new(readyState);
+        }
+        catch (JsonException ex)
+        {
+            throw new DistributedApplicationException("Failed to parse JSON output from 'vercel inspect'.", ex);
+        }
+    }
+
+    public static bool TryGetCliVersion(string output, [NotNullWhen(true)] out Version? version)
+    {
+        // The CLI can print banners/warnings around the version. Extract the first semantic
+        // x.y.z token instead of requiring a line to be exactly the version string.
+        var match = Regex.Match(output, @"(?<!\d)(?<major>\d+)\.(?<minor>\d+)\.(?<patch>\d+)(?!\d)", RegexOptions.CultureInvariant, TimeSpan.FromMilliseconds(100));
+        if (!match.Success)
+        {
+            version = null;
+            return false;
+        }
+
+        version = new(
+            int.Parse(match.Groups["major"].Value, CultureInfo.InvariantCulture),
+            int.Parse(match.Groups["minor"].Value, CultureInfo.InvariantCulture),
+            int.Parse(match.Groups["patch"].Value, CultureInfo.InvariantCulture));
+        return true;
+    }
+
+    public static bool ProjectListContainsProject(string standardOutput, string projectName)
+    {
+        try
+        {
+            using var document = JsonDocument.Parse(standardOutput);
+            foreach (var project in VercelJson.EnumerateArrayOrNamedArray(document.RootElement, "projects"))
+            {
+                if (string.Equals(VercelJson.GetStringProperty(project, "name"), projectName, StringComparison.Ordinal))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+        catch (JsonException ex)
+        {
+            throw new DistributedApplicationException("Failed to parse JSON output from 'vercel project ls'.", ex);
+        }
+    }
+
+    public static bool EnvironmentVariableListContainsName(string standardOutput, string name)
+    {
+        try
+        {
+            using var document = JsonDocument.Parse(standardOutput);
+            foreach (var environmentVariable in VercelJson.EnumerateArrayOrNamedArray(document.RootElement, "envs"))
+            {
+                if (string.Equals(VercelJson.GetStringProperty(environmentVariable, "key"), name, StringComparison.Ordinal)
+                    && string.IsNullOrWhiteSpace(VercelJson.GetStringProperty(environmentVariable, "gitBranch")))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+        catch (JsonException ex)
+        {
+            throw new DistributedApplicationException("Failed to parse JSON output from 'vercel env ls'.", ex);
+        }
+    }
+
+    public static string GetTrimmedOutput(string output)
+        => string.IsNullOrWhiteSpace(output) ? "<empty>" : output.Trim();
+
+    private static VercelDeploymentResult? TryGetJsonDeploymentResult(string standardOutput)
+    {
+        if (!standardOutput.AsSpan().TrimStart().StartsWith("{", StringComparison.Ordinal))
+        {
+            return null;
+        }
+
+        try
+        {
+            using var document = JsonDocument.Parse(standardOutput);
+            var root = document.RootElement;
+
+            if (TryGetDeploymentResult(root, out var deploymentResult))
+            {
+                return deploymentResult;
+            }
+        }
+        catch (JsonException)
+        {
+            // Older Vercel CLI output is plain text; fall back to line-based URL extraction.
+        }
+
+        return null;
+    }
+
+    private static bool TryGetDeploymentResult(JsonElement root, [NotNullWhen(true)] out VercelDeploymentResult? deploymentResult)
+    {
+        // Parse the Vercel deploy JSON shapes observed from different CLI versions:
+        //   { "deployment": { "url": "https://...", "id": "..." } }
+        //   { "url": "https://...", "id": "..." }
+        // Callers fall back to line-based extraction when deploy output is plain text.
+        if (root.TryGetProperty("deployment", out var deployment)
+            && deployment.TryGetProperty("url", out var nestedUrl)
+            && TryGetHttpUrl(nestedUrl, out var nestedDeploymentUrl))
+        {
+            string? deploymentId = deployment.TryGetProperty("id", out var nestedId) && nestedId.ValueKind == JsonValueKind.String
+                ? nestedId.GetString()
+                : null;
+
+            deploymentResult = new(deploymentId, nestedDeploymentUrl);
+            return true;
+        }
+
+        if (root.TryGetProperty("url", out var url)
+            && TryGetHttpUrl(url, out var rootDeploymentUrl))
+        {
+            string? deploymentId = root.TryGetProperty("id", out var id) && id.ValueKind == JsonValueKind.String
+                ? id.GetString()
+                : null;
+
+            deploymentResult = new(deploymentId, rootDeploymentUrl);
+            return true;
+        }
+
+        deploymentResult = null;
+        return false;
+    }
+
+    private static bool TryGetHttpUrl(JsonElement urlElement, [NotNullWhen(true)] out string? url)
+    {
+        url = urlElement.ValueKind == JsonValueKind.String
+            ? urlElement.GetString()
+            : null;
+
+        return IsHttpUrl(url);
+    }
+
+    private static bool IsHttpUrl([NotNullWhen(true)] string? url)
+        => Uri.TryCreate(url, UriKind.Absolute, out var uri) && uri.Scheme is "https" or "http";
+}
