+++ src/CommunityToolkit.Aspire.Hosting.Vercel/VercelDeploymentPaths.cs
@@ -1,0 +1,7 @@
+namespace CommunityToolkit.Aspire.Hosting.Vercel;
+
+internal static class VercelDeploymentPaths
+{
+    public static string GetDeployDirectory(VercelDeploymentEntry entry)
+        => string.IsNullOrWhiteSpace(entry.DeployDirectory) ? entry.SourceRoot : entry.DeployDirectory;
+}
