+++ src/CommunityToolkit.Aspire.Hosting.Vercel/VercelDeploymentStep.State.cs
@@ -1,25 +1,12 @@
 #pragma warning disable ASPIREPIPELINES001
 #pragma warning disable ASPIREPIPELINES002
-#pragma warning disable ASPIREPIPELINES003
 #pragma warning disable ASPIREPIPELINES004
-#pragma warning disable ASPIRECOMPUTE003
-#pragma warning disable ASPIREPROBES001
 #pragma warning disable CTASPIREVERCEL001
 
-using Aspire.Hosting;
 using Aspire.Hosting.ApplicationModel;
-using Aspire.Hosting.ApplicationModel.Docker;
 using Aspire.Hosting.Pipelines;
-using Aspire.Hosting.Publishing;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
-using System.Net.Sockets;
-using System.Text;
 using System.Text.Json;
-using System.Text.Json.Nodes;
-using System.Text.RegularExpressions;
 
 namespace CommunityToolkit.Aspire.Hosting.Vercel;
 
@@ -29,8 +16,8 @@ internal static partial class VercelDeploymentStep
     {
         var options = environment.GetVercelOptions();
         var stateManager = context.Services.GetRequiredService<IDeploymentStateManager>();
-        var stateSection = await stateManager.AcquireSectionAsync(GetStateSectionName(environment), context.CancellationToken).ConfigureAwait(false);
-        var state = ReadDeploymentState(stateSection);
+        var stateSection = await stateManager.AcquireSectionAsync(VercelDeploymentStateStore.GetSectionName(environment), context.CancellationToken).ConfigureAwait(false);
+        var state = VercelDeploymentStateStore.Read(stateSection);
         if (state is null)
         {
             // Keep no-op destroy cheap and offline. If Aspire has no deployment state,
@@ -39,7 +26,7 @@ internal static partial class VercelDeploymentStep
             return;
         }
 
-        ValidateDeploymentState(environment, options, state);
+        VercelDeploymentStateStore.Validate(environment, options, state);
 
         // Destroy is state-first rather than model-first. The current AppHost may no
         // longer contain the resources that created these Vercel projects, but persisted
@@ -78,7 +65,7 @@ internal static partial class VercelDeploymentStep
                 options,
                 environment,
                 deployment,
-                GetVercelProjectEnvironmentName(state)).ConfigureAwait(false);
+                VercelProjectEnvironment.GetName(state)).ConfigureAwait(false);
         }
 
         if (projects.Length == 0)
@@ -96,7 +83,7 @@ internal static partial class VercelDeploymentStep
             }
             else
             {
-                string[] arguments = BuildDestroyProjectArguments(options, projectName);
+                string[] arguments = VercelCliArguments.BuildDestroyProjectArguments(options, projectName);
                 var result = await runner.RunAsync(VercelCliFileName, arguments, workingDirectory: null, context.CancellationToken, standardInput: "y\n").ConfigureAwait(false);
 
                 if (!result.Succeeded)
@@ -114,7 +101,7 @@ internal static partial class VercelDeploymentStep
                 }
             }
 
-            state = RemoveManagedProjectFromDeploymentState(state, projectName);
+            state = VercelDeploymentStateStore.RemoveManagedProject(state, projectName);
             // Save after each project removal so a later CLI failure leaves retryable state
             // for projects that still exist instead of forgetting partially cleaned resources.
             stateSection.SetValue(JsonSerializer.Serialize(state, JsonOptions));
@@ -123,171 +110,6 @@ internal static partial class VercelDeploymentStep
 
         await stateManager.DeleteSectionAsync(stateSection, context.CancellationToken).ConfigureAwait(false);
     }
-
-    private static async Task ValidateExistingDeploymentStateAsync(
-        PipelineStepContext context,
-        VercelEnvironmentResource environment,
-        VercelEnvironmentOptionsAnnotation options)
-    {
-        var stateManager = context.Services.GetRequiredService<IDeploymentStateManager>();
-        var stateSection = await stateManager.AcquireSectionAsync(GetStateSectionName(environment), context.CancellationToken).ConfigureAwait(false);
-        var existingState = ReadDeploymentState(stateSection);
-        if (existingState is not null)
-        {
-            ValidateDeploymentState(environment, options, existingState);
-        }
-    }
-
-    private static async Task<PreviousVercelDeployment?> GetPreviousDeploymentStateEntryAsync(
-        PipelineStepContext context,
-        VercelEnvironmentResource environment,
-        VercelEnvironmentOptionsAnnotation options,
-        string resourceName,
-        string projectName)
-    {
-        var stateManager = context.Services.GetRequiredService<IDeploymentStateManager>();
-        var stateSection = await stateManager.AcquireSectionAsync(GetStateSectionName(environment), context.CancellationToken).ConfigureAwait(false);
-        var existingState = ReadDeploymentState(stateSection);
-        if (existingState is null)
-        {
-            return null;
-        }
-
-        ValidateDeploymentState(environment, options, existingState);
-        var entry = existingState.Deployments.FirstOrDefault(deployment =>
-            string.Equals(deployment.ResourceName, resourceName, StringComparison.Ordinal)
-            && string.Equals(deployment.ProjectName, projectName, StringComparison.Ordinal));
-
-        return entry is null
-            ? null
-            : new(entry, GetVercelProjectEnvironmentName(existingState));
-    }
-
-    private static async Task SaveDeploymentStateEntryAsync(
-        PipelineStepContext context,
-        VercelEnvironmentResource environment,
-        VercelEnvironmentOptionsAnnotation options,
-        VercelDeploymentStateEntry deployment)
-    {
-        var stateManager = context.Services.GetRequiredService<IDeploymentStateManager>();
-        var stateSection = await stateManager.AcquireSectionAsync(GetStateSectionName(environment), context.CancellationToken).ConfigureAwait(false);
-        var existingState = ReadDeploymentState(stateSection);
-        var state = existingState is null
-            ? CreateDeploymentState(environment, options, [deployment])
-            : MergeDeploymentState(environment, options, existingState, deployment);
-
-        stateSection.SetValue(JsonSerializer.Serialize(state, JsonOptions));
-
-        await stateManager.SaveSectionAsync(stateSection, context.CancellationToken).ConfigureAwait(false);
-    }
-
-    private static VercelDeploymentState CreateDeploymentState(
-        VercelEnvironmentResource environment,
-        VercelEnvironmentOptionsAnnotation options,
-        VercelDeploymentStateEntry[] deployments)
-        => new(
-            DeploymentStateSchemaVersion,
-            environment.Name,
-            NormalizeScope(options.Scope),
-            NormalizeTarget(options.Target),
-            options.Production,
-            deployments);
-
-    private static VercelDeploymentState MergeDeploymentState(
-        VercelEnvironmentResource environment,
-        VercelEnvironmentOptionsAnnotation options,
-        VercelDeploymentState existingState,
-        VercelDeploymentStateEntry deployment)
-    {
-        ValidateDeploymentState(environment, options, existingState);
-
-        return CreateDeploymentState(
-            environment,
-            options,
-            [
-                .. existingState.Deployments.Where(existing =>
-                    !string.Equals(existing.ResourceName, deployment.ResourceName, StringComparison.Ordinal)
-                    || !string.Equals(existing.ProjectName, deployment.ProjectName, StringComparison.Ordinal)),
-                deployment
-            ]);
-    }
-
-    private static VercelDeploymentState? ReadDeploymentState(DeploymentStateSection stateSection)
-    {
-        // DeploymentStateSection storage shape has changed across Aspire builds. Accept the
-        // known wrappers so destroy can still clean up projects created by an older CLI.
-        if (stateSection.Data.TryGetPropertyValue("value", out JsonNode? value)
-            && value is not null)
-        {
-            return DeserializeDeploymentState(value);
-        }
-
-        value = stateSection.Data.FirstOrDefault().Value;
-        if (value is not null)
-        {
-            return DeserializeDeploymentState(value);
-        }
-
-        if (stateSection.Data.ContainsKey("schemaVersion"))
-        {
-            return stateSection.Data.Deserialize<VercelDeploymentState>(JsonOptions);
-        }
-
-        return null;
-    }
-
-    private static VercelDeploymentState? DeserializeDeploymentState(JsonNode value)
-    {
-        return value.GetValueKind() == JsonValueKind.String
-            ? JsonSerializer.Deserialize<VercelDeploymentState>(value.GetValue<string>(), JsonOptions)
-            : value.Deserialize<VercelDeploymentState>(JsonOptions);
-    }
-
-    private static void ValidateDeploymentState(
-        VercelEnvironmentResource environment,
-        VercelEnvironmentOptionsAnnotation options,
-        VercelDeploymentState state)
-    {
-        if (state.SchemaVersion != DeploymentStateSchemaVersion)
-        {
-            throw new DistributedApplicationException($"Vercel deployment state for environment '{environment.Name}' uses unsupported schema version '{state.SchemaVersion}'. Redeploy the environment before destroying it.");
-        }
-
-        if (!string.Equals(state.Environment, environment.Name, StringComparison.Ordinal))
-        {
-            throw new DistributedApplicationException($"Vercel deployment state for environment '{state.Environment}' cannot be used to destroy environment '{environment.Name}'.");
-        }
-
-        string? configuredScope = NormalizeScope(options.Scope);
-        if (!string.Equals(state.Scope, configuredScope, StringComparison.Ordinal))
-        {
-            string stateScope = string.IsNullOrWhiteSpace(state.Scope) ? "<default>" : state.Scope;
-            string requestedScope = string.IsNullOrWhiteSpace(configuredScope) ? "<default>" : configuredScope;
-            throw new DistributedApplicationException($"Vercel deployment state for environment '{environment.Name}' was created for scope '{stateScope}', but destroy is configured for scope '{requestedScope}'. Use the same Vercel scope that created the deployment state.");
-        }
-    }
-
-    private static string? NormalizeScope(string? scope)
-        => string.IsNullOrWhiteSpace(scope) ? null : scope;
-
-    private static string? NormalizeTarget(string? target)
-        => string.IsNullOrWhiteSpace(target) ? null : target;
-
-    private static string? GetProductionUrl(VercelEnvironmentOptionsAnnotation options, string projectName)
-        => options.Production ? $"https://{projectName}.vercel.app" : null;
-
-    private static string GetDeployDirectory(VercelDeploymentEntry entry)
-        => string.IsNullOrWhiteSpace(entry.DeployDirectory) ? entry.SourceRoot : entry.DeployDirectory;
-
-    private static VercelDeploymentState RemoveManagedProjectFromDeploymentState(VercelDeploymentState state, string projectName)
-        => state with
-        {
-            Deployments = state.Deployments
-                .Where(deployment => !deployment.ManagedByAspire || !string.Equals(deployment.ProjectName, projectName, StringComparison.Ordinal))
-                .ToArray()
-        };
-
-    private static string GetStateSectionName(VercelEnvironmentResource environment) => $"{StateSectionNamePrefix}{environment.Name}";
 
     private static async Task<bool> ProjectExistsAsync(
         PipelineStepContext context,
@@ -298,14 +120,14 @@ internal static partial class VercelDeploymentStep
         // Avoid treating localized or reformatted CLI errors as provider state. The Vercel
         // CLI exposes project lists as JSON, so destroy checks exact project names before
         // deleting and again after a failed delete to distinguish races from real failures.
-        string[] arguments = BuildListProjectsArguments(options, projectName);
+        string[] arguments = VercelCliArguments.BuildListProjectsArguments(options, projectName);
         var result = await runner.RunAsync(VercelCliFileName, arguments, workingDirectory: null, context.CancellationToken).ConfigureAwait(false);
         if (!result.Succeeded)
         {
             throw CreateCliException($"list Vercel projects while checking for '{projectName}'", VercelCliFileName, result);
         }
 
-        return ProjectListContainsProject(result.StandardOutput, projectName);
+        return VercelCliOutputParser.ProjectListContainsProject(result.StandardOutput, projectName);
     }
 
     private static async Task<bool> ProjectEnvironmentVariableExistsAsync(
@@ -319,14 +141,14 @@ internal static partial class VercelDeploymentStep
         // `vercel env rm` reports absence as human text, but `vercel env ls --format=json`
         // returns the linked project's exact keys. Use that provider read for idempotent
         // stale-secret cleanup instead of parsing failure prose.
-        string[] arguments = BuildListProjectEnvironmentVariablesArguments(options, projectLinkDirectory, targetEnvironment);
+        string[] arguments = VercelCliArguments.BuildListProjectEnvironmentVariablesArguments(options, projectLinkDirectory, targetEnvironment);
         var result = await runner.RunAsync(VercelCliFileName, arguments, projectLinkDirectory, context.CancellationToken).ConfigureAwait(false);
         if (!result.Succeeded)
         {
             throw CreateCliException($"list Vercel project environment variables before removing '{name}'", VercelCliFileName, result);
         }
 
-        return EnvironmentVariableListContainsName(result.StandardOutput, name);
+        return VercelCliOutputParser.EnvironmentVariableListContainsName(result.StandardOutput, name);
     }
 
     private static async Task ConfigureProjectEnvironmentVariablesAsync(
@@ -364,10 +186,10 @@ internal static partial class VercelDeploymentStep
             return;
         }
 
-        string targetEnvironment = GetVercelProjectEnvironmentName(options);
+        string targetEnvironment = VercelProjectEnvironment.GetName(options);
         foreach (var environmentVariable in environmentVariables.OrderBy(static variable => variable.Key, StringComparer.Ordinal))
         {
-            string[] arguments = BuildAddProjectEnvironmentVariableArguments(options, projectLinkDirectory, environmentVariable.Key, targetEnvironment);
+            string[] arguments = VercelCliArguments.BuildAddProjectEnvironmentVariableArguments(options, projectLinkDirectory, environmentVariable.Key, targetEnvironment);
             var result = await runner.RunAsync(VercelCliFileName, arguments, projectLinkDirectory, context.CancellationToken, standardInput: environmentVariable.Value).ConfigureAwait(false);
             if (!result.Succeeded)
             {
@@ -392,7 +214,7 @@ internal static partial class VercelDeploymentStep
                 continue;
             }
 
-            string[] arguments = BuildRemoveProjectEnvironmentVariableArguments(options, projectLinkDirectory, name, targetEnvironment);
+            string[] arguments = VercelCliArguments.BuildRemoveProjectEnvironmentVariableArguments(options, projectLinkDirectory, name, targetEnvironment);
             var result = await runner.RunAsync(VercelCliFileName, arguments, projectLinkDirectory, context.CancellationToken).ConfigureAwait(false);
             if (!result.Succeeded
                 && await ProjectEnvironmentVariableExistsAsync(context, runner, options, projectLinkDirectory, name, targetEnvironment).ConfigureAwait(false))
@@ -412,12 +234,12 @@ internal static partial class VercelDeploymentStep
     {
         var outputService = context.Services.GetRequiredService<IPipelineOutputService>();
         string projectLinkDirectory = Path.Combine(outputService.GetTempDirectory(environment), ".vercel-projects", deployment.ProjectName);
-        DeleteDirectoryIfExists(projectLinkDirectory);
+        VercelFileSystem.DeleteDirectoryIfExists(projectLinkDirectory);
         Directory.CreateDirectory(projectLinkDirectory);
 
         try
         {
-            string[] linkArguments = BuildLinkProjectArguments(options, projectLinkDirectory, deployment.ProjectId ?? deployment.ProjectName);
+            string[] linkArguments = VercelCliArguments.BuildLinkProjectArguments(options, projectLinkDirectory, deployment.ProjectId ?? deployment.ProjectName);
             var linkResult = await runner.RunAsync(VercelCliFileName, linkArguments, projectLinkDirectory, context.CancellationToken).ConfigureAwait(false);
             if (!linkResult.Succeeded)
             {
@@ -435,7 +257,7 @@ internal static partial class VercelDeploymentStep
         }
         finally
         {
-            DeleteDirectoryIfExists(projectLinkDirectory);
+            VercelFileSystem.DeleteDirectoryIfExists(projectLinkDirectory);
         }
     }
 }
