+++ src/CommunityToolkit.Aspire.Hosting.Vercel/VercelFileSystem.cs
@@ -1,0 +1,20 @@
+namespace CommunityToolkit.Aspire.Hosting.Vercel;
+
+internal static class VercelFileSystem
+{
+    public static void DeleteFileIfExists(string path)
+    {
+        if (File.Exists(path))
+        {
+            File.Delete(path);
+        }
+    }
+
+    public static void DeleteDirectoryIfExists(string path)
+    {
+        if (Directory.Exists(path))
+        {
+            Directory.Delete(path, recursive: true);
+        }
+    }
+}
